requirements.txt
affine==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
ansicon==1.89.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
basemap
basemap-data==1.3.2
basemap-data-hires==1.3.2
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.3
bokeh==3.2.2
boto3==1.28.44
botocore==1.31.44
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contextily==1.4.0
contourpy==1.1.0
croniter==1.3.15
cycler==0.11.0
dask==2023.9.1
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distributed==2023.9.1
docopt==0.6.2
dtlpy==1.83.10
earthengine-api==0.1.362
efficientnet-pytorch==0.7.1
einops==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.18.0
filelock==3.12.2
Fiona==1.9.4.post1
Flask==3.0.0
folium==0.13.0
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
fuzzyfinder==2.1.0
gcloud==0.18.3
gcsfs==2023.6.0
GDAL==3.7.2
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
geos==0.2.3
gitdb==4.0.10
GitPython==3.1.35
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
inquirer==3.1.3
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jinxed==1.2.0
jmespath==1.0.1
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
kornia==0.6.12
libpysal==4.7.0
lightning==1.9.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
locket==1.0.0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mercantile==1.2.1
mistune==3.0.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
munch==4.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
ordered-set==4.1.0
overrides==7.4.0
OWSLib==0.29.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.2
patool==1.15.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.10.0
pretrainedmodels==0.7.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pyproj==3.6.0
pyshp==2.3.1
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rasterio==1.3.8
rasterstats==0.19.0
rastertodataframe==0.2.1
readchar==4.0.5
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
Rtree==1.0.1
s3transfer==0.6.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.2
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
statsmodels==0.14.0
sympy==1.12
tabulate==0.8.9
tblib==2.0.0
terminado==0.17.1
threadpoolctl==3.2.0
timm==0.6.12
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchgeo==0.4.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
utm==0.7.0
uvicorn==0.23.2
validators==0.18.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
webvtt-py==0.4.3
Werkzeug==3.0.1
widgetsnbextension==4.0.8
xyzservices==2023.7.0
yarg==0.1.9
yarl==1.9.2
zict==3.0.0
zipp==3.16.2
